\name{CodeinLogMetabolitesUrine}
\alias{CodeinLogMetabolitesUrine}

\docType{data}
\title{Codeine metabolite concentrations in urine.}

\description{Data set from a pharmacogenetic investigation that assessed the formation of morphine from codeine in the presence of variants in cytochrome P450 2D6.}

\usage{data("CodeinLogMetabolitesUrine")}

\details{
Size 50 x 5 , stored in \code{CodeinLogMetabolitesUrine["MOR", "M3G", "M6G", "COD", "C6G"]}
}

\examples{
data(CodeinLogMetabolitesUrine)
str(CodeinLogMetabolitesUrine)
}
