\name{LipidsPsychiatricPat}
\alias{LipidsPsychiatricPat}

\docType{data}
\title{Chromatography mass spectrometry of lipid mediators measured in blood samples.}

\description{Data set from a lipidomic investigation that assessed lipid mediators in blood samples in patients.}

\usage{data("LipidsPsychiatricPat")}

\details{
Size 94 x 8 , stored in \code{LipidsPsychiatricPat["S1P", "C16Sphinganin", "C16Cer", "C20Cer", "C24Cer", "C24_1Cer", "C16GluCer", "C16LacCer"]}
}

\examples{
data(LipidsPsychiatricPat)
str(LipidsPsychiatricPat)
}
