\name{PainThresholds}
\alias{PainThresholds}

\docType{data}
\title{Psychophysical data from an investigation of pain thresholds.}

\description{Data set from an investigation of pain thresholds to various stimuli in healthy volunteers.}

\usage{data("PainThresholds")}

\details{
Size 125 x 8 , stored in \code{PainThresholds["von.Frey", "von.Frey.Caps", "Heat", "Heat.Caps", 
"Cold", "Cold.Menth", "Pressure", "Electric"]}
}

\examples{
data(PainThresholds)
str(PainThresholds)
}

