\name{QSTpainEJPtransf}
\alias{QSTpainEJPtransf}

\docType{data}
\title{Psychophysical data from a clinical quantitative sensory testing study.}

\description{Data set from a psychophysical investigation in a clinical quantitative sensory testing study in healthy subjects.}

\usage{data("QSTpainEJPtransf")}

\details{
Size 72 x 19 , stored in \code{QSTpainEJPtransf["PressureThr", "PressureTol", "TSACold", "ElectricThr", "ElectricTol", "Co2Thr", 
"CO2VAS", "LaserThr", "LaserVAS", "CDT", "WDT", "TSL", "CPT", "HPT", "PPT", "MPT", "MPS", "WUR", "MDT"]}
}

\examples{
data(QSTpainEJPtransf)
str(QSTpainEJPtransf)
}

